+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # Changelog
 
-## Unreleased
+## 3.2.0 - 2023-12-06
 
 ### Added
 - Added an icon to link to embedded asset URLs when an embedded asset is selected in an assets field (inspired by the Embedded Assets Link plugin for Craft 3)